+++ src/app/services/api.ts
@@ -84,10 +84,6 @@ const ldoBaseQuery =
         }
       })
 
-      console.log(ldo)
-
-      console.log(Object.getPrototypeOf(Object.getPrototypeOf(ldo)))
-
       await fetch(documentUrl, {
         method: 'PATCH',
         body: await ldo.$toSparqlUpdate(),
