+++ src/components/AccommodationView/AccommodationView.module.scss
