+++ src/config/leaflet.ts
@@ -1,0 +1,16 @@
+import 'leaflet'
+import * as L from 'leaflet'
+import icon2 from 'leaflet/dist/images/marker-icon-2x.png'
+import icon from 'leaflet/dist/images/marker-icon.png'
+import iconShadow from 'leaflet/dist/images/marker-shadow.png'
+
+// @ts-ignore
+delete L.Icon.Default.prototype._getIconUrl
+L.Icon.Default.mergeOptions({
+  iconUrl: icon,
+  iconRetinaUrl: icon2,
+  shadowUrl: iconShadow,
+})
+
+export const tileServer =
+  'https://{s}.tile-cyclosm.openstreetmap.fr/cyclosm/{z}/{x}/{y}.png'
