+++ src/pages/MyOffers.tsx
@@ -1,8 +1,8 @@
 import { skipToken } from '@reduxjs/toolkit/dist/query'
 import { comunicaApi } from 'app/services/comunicaApi'
 import { Button, Loading } from 'components'
-import { Accommodation as AccommodationView } from 'components/Accommodation/Accommodation'
 import { AccommodationForm } from 'components/AccommodationForm/AccommodationForm'
+import { AccommodationView } from 'components/AccommodationView/AccommodationView'
 import { useAuth } from 'hooks/useAuth'
 import { usePersonalHospexDocuments } from 'hooks/usePersonalHospexDocuments'
 import { useState } from 'react'
